+++ Code/Client/users/src/routes/Routes.js
@@ -1,5 +1,6 @@
 import React, { useState, useEffect, Profiler } from "react";
 import { Routes, Route, Navigate, useLocation } from "react-router-dom";
+import './route.css'
 import Home from "../pages/Home.js";
 import Signup from "../pages/Signup";
 import Login from "../pages/Login";
@@ -10,11 +11,10 @@ import NoRoute from "../pages/NoRoute.js";
 import { useDispatch } from "react-redux";
 import { useSelector } from "react-redux";
 import { setSignedIn } from "../redux/containers/auth/actions";
-import './route.css'
-import LandingPage from "../pages/carRental/LandingPage.js";
-import ViewCars from "../pages/carRental/showCarsPage/ViewCars.js"
-import ExtendedDetailPage from "../pages/carRental/addOnPage/ExtendedDetailPage.js"
-import RentalBookingPage from "../pages/carRental/bookingPage/RentalBookingPage.js"
+import LandingPage from "../pages/car-rental/LandingPage.js";
+import ViewCars from "../pages/car-rental/showCarsPage/ViewCars.js"
+import ExtendedDetailPage from "../pages/car-rental/addOnPage/ExtendedDetailPage.js"
+import RentalBookingPage from "../pages/car-rental/bookingPage/RentalBookingPage.js"
 import BlogsNews from "../pages/blogs&news/BlogsNews.js";
 import BlogsDetails from "../pages/blogs&news/BlogsDetails.js";
 import CareerPage from "../pages/career/CareerPage.js";
